Adaptions to allow compilation with AIX compilers   (in particular, va_copy not available)

+++ LinAlgImpl/Serial/IpDenseVector.cpp
@@ -160,7 +160,7 @@ namespace Ipopt
   {
     DBG_ASSERT(initialized_);
     if (homogeneous_) {
-      return sqrt(Dim()) * fabs(scalar_);
+      return sqrt((double)Dim()) * fabs(scalar_);
     }
     else {
       return IpBlasDnrm2(Dim(), values_, 1);
